+++ test/view-configs/axis.js
@@ -1,0 +1,66 @@
+const viewconfig = {
+  editable: true,
+  views: [
+    {
+      uid: 'Cs0jaHTuQXuibqx36Ew1xg',
+      tracks: {
+        top: [
+          {
+            name: 'wgEncodeUwDnaseGm12865RawRep1.hitile',
+            created: '2017-09-11T21:41:54.166605Z',
+            server: 'http://higlass.io/api/v1',
+            tilesetUid: 'VX4h-WzlSgu2VI88Fg5YGw',
+            uid: 'frcXuRouRpa_XSm5awtt3Q',
+            type: 'horizontal-bar',
+            options: {
+              labelColor: 'black',
+              labelPosition: 'topLeft',
+              axisPositionHorizontal: 'right',
+              axisLabelFormatting: 'normal',
+              axisMargin: 10,
+              lineStrokeColor: 'blue',
+              lineStrokeWidth: 1,
+              valueScaling: 'linear',
+              trackBorderWidth: 0,
+              trackBorderColor: 'black',
+              labelTextOpacity: 0.4,
+              showMousePosition: false,
+              mousePositionColor: '#999999',
+              showTooltip: false,
+              name: 'wgEncodeUwDnaseGm12865RawRep1.hitile',
+              align: 'bottom',
+              barFillColor: 'darkgreen',
+              barOpacity: 1
+            },
+            width: 275,
+            height: 220,
+            position: 'top'
+          }
+        ],
+        center: [],
+        left: [],
+        right: [],
+        bottom: [],
+        whole: [],
+        gallery: []
+      },
+      layout: {
+        w: 6,
+        h: 6,
+        x: 0,
+        y: 0,
+        i: 'Cs0jaHTuQXuibqx36Ew1xg',
+        moved: false,
+        static: false
+      },
+      initialXDomain: [1081899806.8550227, 1097764936.0057092],
+      initialYDomain: [-3684650483.7164283, -3642881925.5160756]
+    },
+  ],
+  trackSourceServers: ['http://higlass.io/api/v1'],
+  locationLocks: { locksByViewUid: {}, locksDict: {} },
+  zoomLocks: { locksByViewUid: {}, locksDict: {} },
+  exportViewUrl: 'http://higlass.io/api/v1/viewconfs',
+  valueScaleLocks: { locksByViewUid: {}, locksDict: {} }
+};
+export default viewconfig;